requirements.txt
Django==2.2.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-mdeditor==0.1.14
django-widget-tweaks==1.4.5
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
